# internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: futureEnvirons API
  description: |
    Go backend for the futureEnvirons IoT platform.

    All successful responses use a standard envelope:

    ```json
    { "success": true, "data": { ... } }
    ```

    Errors use:

    ```json
    { "success": false, "error": { "code": "...", "message": "..." } }
    ```
  version: 1.0.0
  contact:
    name: futureEnvirons

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Health
    description: Service health probes
  - name: Auth
    description: Authentication
  - name: Users
    description: User management (SUPER_ADMIN and ORG_ADMIN)

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      description: Load balancer / ops probe (outside the versioned API prefix).
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/HealthData"

  /api/v1/health:
    get:
      tags: [Health]
      summary: Versioned health check
      operationId: getV1Health
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/HealthData"

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: |
        Authenticates a user with email and password. Password verification uses bcrypt in Go.
        User details are loaded from `identity.sp_get_login_details`.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email_id: admin@futureenvirons.com
              password: Admin@123
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                error:
                  code: UNAUTHORIZED
                  message: invalid email or password
        "403":
          description: Account inactive, not approved, or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      description: |
        Validates the access token. Server-side token invalidation will be added when
        `identity.refresh_session` exists. Clients should discard the token locally.
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LogoutResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      description: SUPER_ADMIN sees all users. ORG_ADMIN sees users in their organization only.
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Users]
      summary: Create user
      description: |
        Creates `identity.users` via `identity.sp_manage_user`, then inserts
        `user_credentials` and `user_role_assignment` in the same transaction.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUserByID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Users]
      summary: Soft delete user
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User soft-deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DeleteUserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token returned by the login endpoint.

  schemas:
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object

    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: "#/components/schemas/APIError"

    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
        message:
          type: string

    HealthData:
      type: object
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: futureEnvironsBE
        env:
          type: string
          example: development
        timestamp:
          type: string
          format: date-time

    LoginRequest:
      type: object
      required: [email_id, password]
      properties:
        email_id:
          type: string
          format: email
        password:
          type: string
          format: password

    Role:
      type: object
      properties:
        role_id:
          type: string
          format: uuid
        role_code:
          type: string
        role_name:
          type: string

    UserProfile:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        email_id:
          type: string
          format: email
        contact_no:
          type: string
        organization_id:
          type: string
          format: uuid
        organization_name:
          type: string
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          format: int64
          description: Token lifetime in seconds.
        user:
          $ref: "#/components/schemas/UserProfile"

    LogoutResponse:
      type: object
      properties:
        message:
          type: string
          example: logged out

    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        contact_no:
          type: string
        email_id:
          type: string
          format: email
        address:
          type: string
        organization_id:
          type: string
          format: uuid
        organization_name:
          type: string
        valid_till:
          type: string
          format: date-time
        is_active:
          type: boolean
        is_approved:
          type: boolean
        is_deleted:
          type: boolean
        created_by:
          type: string
          format: uuid
        created_on:
          type: string
          format: date-time
        modified_by:
          type: string
          format: uuid
        modified_on:
          type: string
          format: date-time
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"

    CreateUserRequest:
      type: object
      required: [user_name, email_id, password, organization_id, role_id]
      properties:
        user_name:
          type: string
        contact_no:
          type: string
        email_id:
          type: string
          format: email
        address:
          type: string
        organization_id:
          type: string
          format: uuid
        valid_till:
          type: string
          format: date-time
        is_active:
          type: boolean
        is_approved:
          type: boolean
        password:
          type: string
          format: password
          minLength: 8
        role_id:
          type: string
          format: uuid

    UpdateUserRequest:
      type: object
      properties:
        user_name:
          type: string
        contact_no:
          type: string
        email_id:
          type: string
          format: email
        address:
          type: string
        organization_id:
          type: string
          format: uuid
        valid_till:
          type: string
          format: date-time
        is_active:
          type: boolean
        is_approved:
          type: boolean
        role_id:
          type: string
          format: uuid

    DeleteUserResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        email_id:
          type: string
          format: email
        is_active:
          type: boolean
        is_deleted:
          type: boolean

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Conflict (e.g. duplicate email)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"